Updated after code review:
Removed logging
Slimmed down the Java code
Moved default scheduler
