--- akka-actor/src/main/java/org/jboss/netty/akka/logging/AbstractInternalLogger.java
@@ -1,88 +1,0 @@
-/*
- * Copyright 2009 Red Hat, Inc.
- *
- * Red Hat licenses this file to you under the Apache License, version 2.0
- * (the "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at:
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package org.jboss.netty.akka.logging;
-
-/**
- * A skeletal implementation of {@link InternalLogger}.  This class implements
- * all methods that have a {@link InternalLogLevel} parameter by default to call
- * specific logger methods such as {@link #info(String)} or {@link #isInfoEnabled()}.
- *
- * @author <a href="http://www.jboss.org/netty/">The Netty Project</a>
- * @author <a href="http://gleamynode.net/">Trustin Lee</a>
- * @version $Rev: 2080 $, $Date: 2010-01-26 18:04:19 +0900 (Tue, 26 Jan 2010) $
- */
-public abstract class AbstractInternalLogger implements InternalLogger {
-
-    /**
-     * Creates a new instance.
-     */
-    protected AbstractInternalLogger() {
-        super();
-    }
-
-    public boolean isEnabled(InternalLogLevel level) {
-        switch (level) {
-        case DEBUG:
-            return isDebugEnabled();
-        case INFO:
-            return isInfoEnabled();
-        case WARN:
-            return isWarnEnabled();
-        case ERROR:
-            return isErrorEnabled();
-        default:
-            throw new Error();
-        }
-    }
-
-    public void log(InternalLogLevel level, String msg, Throwable cause) {
-        switch (level) {
-        case DEBUG:
-            debug(msg, cause);
-            break;
-        case INFO:
-            info(msg, cause);
-            break;
-        case WARN:
-            warn(msg, cause);
-            break;
-        case ERROR:
-            error(msg, cause);
-            break;
-        default:
-            throw new Error();
-        }
-    }
-
-    public void log(InternalLogLevel level, String msg) {
-        switch (level) {
-        case DEBUG:
-            debug(msg);
-            break;
-        case INFO:
-            info(msg);
-            break;
-        case WARN:
-            warn(msg);
-            break;
-        case ERROR:
-            error(msg);
-            break;
-        default:
-            throw new Error();
-        }
-    }
-}
