+++ akka-remote/src/main/scala/akka/remote/Gossiper.scala
@@ -4,7 +4,6 @@
 
 package akka.remote
 
-import akka.AkkaApplication
 import akka.actor._
 import akka.actor.Status._
 import akka.event.Logging
@@ -14,7 +13,6 @@ import akka.remote.RemoteProtocol.RemoteSystemDaemonMessageType._
 
 import java.net.InetSocketAddress
 import java.util.concurrent.atomic.AtomicReference
-import java.util.concurrent.TimeUnit
 import java.security.SecureRandom
 import System.{ currentTimeMillis ⇒ newTimestamp }
 
@@ -109,7 +107,6 @@ class Gossiper(remote: Remote) {
   private val failureDetector = remote.failureDetector
   private val connectionManager = new RemoteConnectionManager(app, remote, Map.empty[InetSocketAddress, ActorRef])
   private val seeds = Set(address) // FIXME read in list of seeds from config
-  private val scheduler = app.scheduler
 
   private val address = app.defaultAddress
   private val nodeFingerprint = address.##
@@ -126,8 +123,8 @@ class Gossiper(remote: Remote) {
 
   {
     // start periodic gossip and cluster scrutinization - default is run them every second with 1/2 second in between
-    scheduler schedule (() ⇒ initateGossip(), initalDelayForGossip.toSeconds, gossipFrequency.toSeconds, timeUnit)
-    scheduler schedule (() ⇒ scrutinize(), initalDelayForGossip.toSeconds, gossipFrequency.toSeconds, timeUnit)
+    app.scheduler schedule (() ⇒ initateGossip(), initalDelayForGossip.toSeconds, gossipFrequency.toSeconds, timeUnit)
+    app.scheduler schedule (() ⇒ scrutinize(), initalDelayForGossip.toSeconds, gossipFrequency.toSeconds, timeUnit)
   }
 
   /**
